Add ULP verification for fp32
Add a verifier to check two results are correct within a certain ULP
tolerance for IEEE-754 32-bit floating point values.

Add a test to check the ULP verifier is correct.

+++ reference_model/src/verify/verify_utils.h
@@ -51,7 +51,7 @@ struct UlpInfo
 {
     UlpInfo() = default;
 
-    float ulp;
+    uint64_t ulp;
 };
 
 /// \brief Dot-product verification meta-data
